secure-healthcare-app/backend/requirements.txt
Flask
cryptography
flask-cors
flask-jwt-extended
